feat: expose create(preset) engine so 3rd-party create-<brand> packages can add plugins
Refactor the scaffolder into a reusable library:

- engine.js: holds the scaffolding logic and exports create({ preset, argv }).
  A preset (MVP: { name?, plugins? }) appends its plugins to the scaffolded
  sandbox.config.json, so downstream packages get the full sandbox plus their
  own plugins. preset.name is used only to print the right `npm create <name>`
  in help/empty-dir messages.
- index.js: now a thin bin that just calls create().
- package.json: add "exports" (".": "./engine.js") and ship engine.js so
  wrapper packages can `import { create } from 'create-wp-local-dev-agent-sandbox'`.
- README: document the 'Build your own npm create command' pattern (depend on
  this package, call create() with a preset) — no fork required.

Verified: base CLI unchanged; preset API appends a .zip plugin to the defaults
(mcp-adapter + root-for-agents stay) and writes valid JSON.
